refactor(integrations): give four small vendors a public API to import from
