+++ integrations/elasticsearch/__init__.py
@@ -1,0 +1,7 @@
+"""Elasticsearch integration."""
+
+from __future__ import annotations
+
+from integrations.elasticsearch.client import ElasticsearchClient, ElasticsearchConfig
+
+__all__ = ["ElasticsearchClient", "ElasticsearchConfig"]
