+++ integrations/google_docs/__init__.py
@@ -1,0 +1,7 @@
+"""Google Docs integration."""
+
+from __future__ import annotations
+
+from integrations.google_docs.client import GoogleDocsClient
+
+__all__ = ["GoogleDocsClient"]
