+++ surfaces/cli/wizard/integration_validators/alerting.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from integrations.alertmanager.client import make_alertmanager_client
 from integrations.betterstack import build_betterstack_config, validate_betterstack_config
-from integrations.opsgenie.client import OpsGenieClient, OpsGenieConfig
+from integrations.opsgenie import OpsGenieClient, OpsGenieConfig
 
 from .shared import IntegrationHealthResult
