+++ tests/shared/test_integrations_api_border.py
@@ -103,14 +103,12 @@ _ALLOWED: dict[str, frozenset[str]] = {
             "integrations.dagster.setup",
             "integrations.datadog.setup",
             "integrations.discord.setup",
-            "integrations.elasticsearch.client",
             "integrations.github.identity",
             "integrations.github.login",
             "integrations.github.mcp",
             "integrations.github.mcp_oauth",
             "integrations.github.setup",
             "integrations.gitlab.setup",
-            "integrations.google_docs.client",
             "integrations.grafana.client",
             "integrations.grafana.setup",
             "integrations.hermes.agent",
@@ -140,7 +138,6 @@ _ALLOWED: dict[str, frozenset[str]] = {
             "integrations.openclaw.setup",
             "integrations.openclaw.validate_openclaw_config",
             "integrations.opensearch.setup",
-            "integrations.opsgenie.client",
             "integrations.pagerduty.setup",
             "integrations.posthog.report_prerequisites",
             "integrations.posthog.setup",
@@ -160,7 +157,6 @@ _ALLOWED: dict[str, frozenset[str]] = {
             "integrations.servicenow.setup",
             "integrations.servicenow.verifier",
             "integrations.slack.setup",
-            "integrations.splunk.client",
             "integrations.telegram.alarms",
             "integrations.telegram.credentials",
             "integrations.telegram.setup",
