+++ tests/shared/test_tool_api_border.py
@@ -1,18 +1,18 @@
-"""Consumers import the tool tier through its API, and the surface only shrinks.
+"""Consumers import the tool packages through their public API, and the surface only shrinks.
 
 ``core/tool/`` (contract, execution, registry port) and ``core/tool_framework/``
-(``@tool``, skill guidance, payload utilities) are one tier to everything above
-them. Three doors are open: ``core.tool`` (the contract — what a tool is, how
-it runs, where it is registered), ``core.tool_framework`` (the ``@tool``
-decorator, planning tags, skill guidance), and ``core.tool_framework.utils``
-(schema builders, MCP readers, availability envelopes). Everything reaching past
-a door is listed below, and these allowlists are the measurement of the seam.
+(``@tool``, skill guidance, payload utilities) are one group to everything above
+them. Three public API modules are the allowed way in: ``core.tool`` (the
+contract — what a tool is, how it runs, where it is registered),
+``core.tool_framework`` (the ``@tool`` decorator, planning tags, skill guidance),
+and ``core.tool_framework.utils`` (schema builders, MCP readers, availability
+envelopes). Any import reaching past those is listed below, and these allowlists
+measure how far off that rule the code still is.
 
-Every consumer is now at zero: they use the tier's doors and nothing behind them.
+Every consumer is now at zero: they use the public API modules and nothing behind them.
 
 Each allowlist is compared exactly in both directions: a new internal import
-fails immediately, and an entry no longer imported must be removed. The seam is
-closed — keep it closed.
+fails immediately, and an entry no longer imported must be removed.
 """
 
 from __future__ import annotations
@@ -27,7 +27,7 @@ from tests.shared.tool_api import TOOL_BORDER
 
 REPO_ROOT = Path(__file__).resolve().parents[2]
 
-#: Internal tool-tier modules each consumer still imports directly.
+#: Internal tool-package modules each consumer still imports directly.
 _ALLOWED: dict[str, frozenset[str]] = {
     "tools": frozenset(),
     "integrations": frozenset(),
@@ -38,7 +38,7 @@ _ALLOWED: dict[str, frozenset[str]] = {
 
 
 @pytest.mark.parametrize("consumer", sorted(_ALLOWED))
-def test_consumer_tool_tier_imports_match_the_allowlist(consumer: str) -> None:
+def test_consumer_tool_package_imports_match_the_allowlist(consumer: str) -> None:
     # Arrange / Act
     imported = TOOL_BORDER.internal_imports_under(
         REPO_ROOT / consumer, exclude_parts=frozenset({"tests"})
@@ -48,8 +48,8 @@ def test_consumer_tool_tier_imports_match_the_allowlist(consumer: str) -> None:
     TOOL_BORDER.assert_matches_allowlist(imported, _ALLOWED[consumer], consumer=f"{consumer}/")
 
 
-def test_bootstrap_reaches_the_tool_tier_only_through_its_api() -> None:
-    """The composition root wires tools together; it must not open the tier up."""
+def test_bootstrap_reaches_the_tool_packages_only_through_their_api() -> None:
+    """The composition root wires tools together; it must not open these packages up."""
     # Arrange / Act
     imported = TOOL_BORDER.internal_imports_under(REPO_ROOT / "bootstrap")
 
@@ -57,7 +57,7 @@ def test_bootstrap_reaches_the_tool_tier_only_through_its_api() -> None:
     assert imported == set()
 
 
-def test_the_contract_door_survives_being_imported_second() -> None:
+def test_the_contract_api_survives_being_imported_second() -> None:
     """``import core.llm.types`` must not re-enter a half-built ``core.tool``.
 
     ``core.tool.execution`` needs ``ToolCall`` from ``core.llm.types``, which in
